#7 bug fixing in the react router (#9)
Functional Components Example Typo:
In the example provided for functional components, there's a typo in the
word "function". It should be corrected for accuracy.

 Class Components Example Missing Import Statement:
The example for class components lacks an import statement for React.
While this might be understood by experienced developers, it could
confuse beginners and should be addressed for consistency and clarity.

 Lack of Explanation on Lifecycle Methods for Class Components:
The explanation for class components doesn't cover lifecycle methods,
which are crucial for understanding component behavior and managing side
effects. Including a brief mention or link to further resources would
enhance the completeness of the explanation.